Create ParallelProcessing.cpp
Solution to "Problem: Parallel processing" in coursera course : Data Structure
https://www.coursera.org/learn/data-structures/programming/Kru9d/programming-assignment-2-priority-queues-and-disjoint-sets

Task:
 You have a program which is parallelized and uses 𝑛 independent threads to process the given list
of 𝑚 jobs. Threads take jobs in the order they are given in the input. If there is a free thread,
it immediately takes the next job from the list. If a thread has started processing a job, it doesn’t
interrupt or stop until it finishes processing the job. If several threads try to take jobs from the list
simultaneously, the thread with smaller index takes the job. For each job you know exactly how long
will it take any thread to process this job, and this time is the same for all the threads. You need to
determine for each job which thread will process it and when will it start processing.

Input Format:
 The first line of the input contains integers 𝑛 and 𝑚.
The second line contains 𝑚 integers 𝑡𝑖 — the times in seconds it takes any thread to process 𝑖-th job.
The times are given in the same order as they are in the list from which threads take jobs.
Threads are indexed starting from 0.
Constraints. 1 ≤ 𝑛 ≤ 105
; 1 ≤ 𝑚 ≤ 105
; 0 ≤ 𝑡𝑖 ≤ 109
.

Output Format:
Output exactly 𝑚 lines. 𝑖-th line (0-based index is used) should contain two spaceseparated
integers — the 0-based index of the thread which will process the 𝑖-th job and the time in
seconds when it will start processing that job.


Solving Idea:
We create a min-heap with the available threads, where the priority is determined by the starting time. 
To assign job, we extract the thread with the minimum priority (min starting time), assign it a job (which updates it's starting time), then add it back to the heap structure. 
We repeat this until all jobs are processed. 

Using std::priority_queue to build a min-heap structure. Tried using my own heap, but it still has some issues. Need to fix my own heap data structure.